requirements.txt
Django==4.2.5
django-environ==0.11.2
django-extensions==3.2.3
Jinja2==3.1.2
pdfkit==1.0.0
pydantic-settings==2.0.3
python-dateutil==2.8.2
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0